Update package version to 1.1.0, add license information, and enhance README with multicolor support details. Remove unused create-hugeicon-singleton component and refactor HugeiconsIcon implementation.
